Add test_fork_reactor, move it with test_fork_safe to a new file
